refactor(mcp-app): use shared plugin builder, migrate to rolldownOptions
Replaces duplicated buildRendererPlugin() with shared builder,
converts build scripts to TypeScript, replaces deprecated
inlineDynamicImports with codeSplitting: false, and rollupOptions
with rolldownOptions.
